Crate chksum_hash_core

Core traits and functions for batch and stream hash computation.

Setup

To use this crate, add the following entry to your Cargo.toml file in the dependencies section:

[dependencies]
chksum-hash-core = "0.0.0"

Alternatively, you can use the cargo add subcommand:

cargo add chksum-hash-core

Example Crates

For implementation-specific examples, refer to the source code of the following crates:

chksum-hash-md5

chksum-hash-sha1

chksum-hash-sha2

chksum-hash-sha2-224

chksum-hash-sha2-256

chksum-hash-sha2-384

chksum-hash-sha2-512

License

This crate is licensed under the MIT License.

Traits

Digest

A trait for hash digests.

Finalize

A trait for hash objects in a finalized state.

Update

A trait for hash objects in an in-progress state.

Functions

default

Creates a default hash.

hash

Computes the hash of the given input.